build: choose .a lib file to copy based on xcode build target

+++ safari/scripts/build_uniffi.sh
@@ -14,5 +14,12 @@ rm -rf $PROJECT_DIR/rust
 # if new swift binding files are added, they need to be added into XCode
 CRATE_DIR="$PROJECT_DIR/../rust"
 cp -RpL $CRATE_DIR/swift/. $PROJECT_DIR/rust
-cp $TARGET_DIR/aarch64-apple-ios/release/libyomikiri_rs.a "$PROJECT_DIR/rust/libyomikiri_rs.a"
-cp $TARGET_DIR/aarch64-apple-ios-sim/release/libyomikiri_rs.a "$PROJECT_DIR/rust/libyomikiri_rs_sim.a"
+
+# Both iphone and ipad simulator
+if [ "$PLATFORM_NAME" = "iphonesimulator" ]; then
+  cp $TARGET_DIR/aarch64-apple-ios-sim/release/libyomikiri_rs.a "$PROJECT_DIR/rust/libyomikiri_rs.a"
+else
+  cp $TARGET_DIR/aarch64-apple-ios/release/libyomikiri_rs.a "$PROJECT_DIR/rust/libyomikiri_rs.a"
+fi
+
+
